Unspent outputs

+++ base_layer/core_storage_postgres/migrations/2020-03-08-104615_create_unspent_outputs/up.sql
@@ -1,7 +1,7 @@
 create table unspent_outputs(
     hash text not null primary key,
     features_flags int not null,
-    features_maturity numeric not null,
+    features_maturity bigint not null,
     commitment text not null,
     proof bytea not null,
     created_at timestamp not null default current_timestamp,
